# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-Brand Shopify E-Commerce & B2B Wholesale Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /super-admin/stores:
    get:
      operationId: listStores
      tags: [super-admin]
      summary: List all storefronts
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [super-admin]
      summary: Create a new storefront
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /super-admin/stores/{storeId}:
    get:
      operationId: getStore
      tags: [super-admin]
      summary: Get a store by ID
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    patch:
      operationId: updateStore
      tags: [super-admin]
      summary: Update a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    delete:
      operationId: deleteStore
      tags: [super-admin]
      summary: Delete a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /super-admin/stores/{storeId}/admins:
    get:
      operationId: listStoreAdmins
      tags: [super-admin]
      summary: List admins for a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createStoreAdmin
      tags: [super-admin]
      summary: Create a store admin account
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /super-admin/stores/{storeId}/admins/{adminId}:
    delete:
      operationId: deleteStoreAdmin
      tags: [super-admin]
      summary: Delete a store admin
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
        - name: adminId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /super-admin/orders/{orderId}/status:
    patch:
      operationId: updateOrderStatusSuperAdmin
      tags: [super-admin]
      summary: Advance order fulfillment status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── STORE ADMIN ────────────────────────────────────────
...
components:
  schemas:
...
    Store:
      type: object
      required: [id, name, slug, isActive, demoMode, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        logoText:
          type: ["string", "null"]
        logoImageUrl:
          type: ["string", "null"]
        announcementBar:
          type: ["string", "null"]
        primaryColor:
          type: string
        accentColor:
          type: string
        fontFamily:
          type: string
          enum: [Inter, Playfair Display, Outfit, Space Grotesk]
        isActive:
          type: boolean
        demoMode:
          type: boolean
        shopifyDomain:
          type: ["string", "null"]
        shopifyStorefrontToken:
          type: ["string", "null"]
        shopifyAdminKey:
          type: ["string", "null"]
        adminCount:
          type: number
        productCount:
          type: number
        orderCount:
          type: number
        totalRevenue:
          type: number
        createdAt:
          type: string
    StoreInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        logoText:
          type: string
        logoImageUrl:
          type: string
        announcementBar:
          type: string
        primaryColor:
          type: string
        accentColor:
          type: string
        fontFamily:
          type: string
          enum: [Inter, Playfair Display, Outfit, Space Grotesk]
        demoMode:
          type: boolean
        shopifyDomain:
          type: string
        shopifyStorefrontToken:
          type: string
        shopifyAdminKey:
          type: string
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
        logoText:
          type: string
        logoImageUrl:
          type: string
        announcementBar:
          type: string
        primaryColor:
          type: string
        accentColor:
          type: string
        fontFamily:
          type: string
          enum: [Inter, Playfair Display, Outfit, Space Grotesk]
        isActive:
          type: boolean
        demoMode:
          type: boolean
        shopifyDomain:
          type: string
        shopifyStorefrontToken:
          type: string
        shopifyAdminKey:
          type: string
...
    AdminUser:
      type: object
      required: [id, email, storeId, storeName, createdAt]
      properties:
        id:
          type: number
        email:
          type: string
        storeId:
          type: number
        storeName:
          type: string
        createdAt:
          type: string
    AdminUserInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    # ── Product ───────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, storeId, type, status, fulfillmentStep, total, paymentTerms, createdAt]
      properties:
        id:
          type: number
        storeId:
          type: number
        storeName:
          type: ["string", "null"]
        type:
          type: string
          enum: [b2b, b2c]
        status:
          type: string
          enum: [received, production, shipped, delivered]
        fulfillmentStep:
          type: number
        total:
          type: number
        paymentTerms:
          type: string
          enum: [cod, net30, card]
        b2bClientId:
          type: ["number", "null"]
        b2bCompanyName:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        shopifyOrderId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderItem:
      type: object
      required: [id, orderId, productId, productName, quantity, unitPrice, lineTotal]
      properties:
        id:
          type: number
        orderId:
          type: number
        productId:
          type: number
        productName:
          type: string
        variantId:
          type: ["number", "null"]
        variantLabel:
          type: ["string", "null"]
        quantity:
          type: number
        unitPrice:
          type: number
        lineTotal:
          type: number
        artworkId:
          type: ["number", "null"]
        artworkName:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
...
    OrderStatusUpdate:
      type: object
      required: [status, fulfillmentStep]
      properties:
        status:
          type: string
          enum: [received, production, shipped, delivered]
        fulfillmentStep:
          type: number

    # ── Analytics ─────────────────────────────────────────────
...