# lib/api-zod/src/generated/types/adminUserInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Brand Shopify E-Commerce & B2B Wholesale Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserInput {
  email: string;
  password: string;
}